mv test folder into archive folder
