+++ README.md
@@ -85,5 +85,4 @@ This project is released under the **MIT License**.
 
 ---
 
-Let me know if you need modifications! 🚀
