--- project/archive/subject_list_temp.txt
@@ -1,3 +1,0 @@
-S0120/S0120_2
-S0135/S0135_1
-S0135/S0135_2
