--- project/archive/subject_list_temp_ori.txt
@@ -1,79 +1,0 @@
-S0113/S0113_1
-S0120/S0120_2
-S0135/S0135_1
-S0135/S0135_2
-S0136/S0136_1
-S0136/S0136_2
-S0139/S0139_1
-S0139/S0139_2
-S0140/S0140_1
-S0140/S0140_2
-S0147/S0147_1
-S0147/S0147_2
-S0148/S0148_1
-S0148/S0148_2
-S0153/S0153_1
-S0153/S0153_2
-S0154/S0154_1
-S0154/S0154_2
-S0155/S0155_1
-S0155/S0155_2
-S0156/S0156_1
-S0156/S0156_2
-S0165/S0165_1
-S0165/S0165_2
-S0167/S0167_1
-S0167/S0167_2
-S0168/S0168_1
-S0168/S0168_2
-S0169/S0169_1
-S0170/S0170_1
-S0170/S0170_2
-S0171/S0171_1
-S0171/S0171_2
-S0172/S0172_1
-S0172/S0172_2
-S0173/S0173_1
-S0173/S0173_2
-S0178/S0178_1
-S0185/S0185_1
-S0185/S0185_2
-S0186/S0186_1
-S0186/S0186_2
-S0188/S0188_1
-S0188/S0188_2
-S0189/S0189_1
-S0189/S0189_2
-S0190/S0190_1
-S0190/S0190_2
-S0191/S0191_1
-S0191/S0191_2
-S0193/S0193_1
-S0193/S0193_2
-S0194/S0194_1
-S0194/S0194_2
-S0196/S0196_1
-S0196/S0196_2
-S0197/S0197_1
-S0197/S0197_2
-S0198/S0198_1
-S0198/S0198_2
-S0199/S0199_1
-S0199/S0199_2
-S0200/S0200_1
-S0200/S0200_2
-S0202/S0202_1
-S0202/S0202_2
-S0204/S0204_1
-S0204/S0204_2
-S0205/S0205_1
-S0205/S0205_2
-S0206/S0206_1
-S0206/S0206_2
-S0207/S0207_1
-S0207/S0207_2
-S0209/S0209_1
-S0209/S0209_2
-S0210/S0210_1
-S0210/S0210_2
-S0213/S0213_2
