+++ project/archive/tests/related_work/arefnezhad2019/ddd_Aref2019.py
@@ -1,0 +1,269 @@
+import warnings
+warnings.filterwarnings("ignore")
+
+# 必要なライブラリのインポート
+import pandas as pd
+import numpy as np
+from sklearn.svm import SVC
+from sklearn.model_selection import train_test_split
+from sklearn.metrics import accuracy_score
+from sklearn.feature_selection import mutual_info_classif
+from scipy.stats import ttest_ind, pearsonr
+from scipy.optimize import differential_evolution
+from pyswarm import pso  # PSOのためのライブラリ
+
+# 特徴量の計算関数
+def calculate_features_for_signals(angle_window, speed_window):
+    features = {}
+    features['angle_range'] = angle_window.max() - angle_window.min()
+    features['angle_standard_dev'] = angle_window.std()
+    features['angle_energy'] = np.sum(angle_window ** 2)
+    features['angle_zcr'] = ((angle_window[:-1] * angle_window[1:]) < 0).sum() / len(angle_window)
+    features['angle_first_quartile'] = np.percentile(angle_window, 25)
+    features['angle_second_quartile'] = np.percentile(angle_window, 50)
+    features['angle_third_quartile'] = np.percentile(angle_window, 75)
+    features['angle_kfd'] = np.log10(len(angle_window)) / (np.log10(len(angle_window)) + np.log10(np.mean(np.abs(np.diff(angle_window)))))
+    features['angle_skewness'] = angle_window.skew()
+    features['angle_kurtosis'] = angle_window.kurtosis()
+    features['angle_sam_entropy'] = -np.sum(np.log2(angle_window / angle_window.sum()) * (angle_window / angle_window.sum()))
+    features['angle_shannon_entropy'] = -np.sum(angle_window * np.log2(angle_window + 1e-10))
+    features['angle_freq_variability'] = np.var(np.fft.fft(angle_window))
+    features['angle_spectral_entropy'] = -np.sum(np.log2(np.abs(np.fft.fft(angle_window)) + 1e-10))
+    features['angle_spectral_flux'] = np.sum(np.abs(np.diff(np.fft.fft(angle_window))))
+    features['angle_center_gravity_freq'] = np.sum(np.fft.fftfreq(len(angle_window)) * np.abs(np.fft.fft(angle_window)))
+    features['angle_dominant_freq'] = np.argmax(np.abs(np.fft.fft(angle_window)))
+    features['angle_avg_psd'] = np.mean(np.abs(np.fft.fft(angle_window))**2)
+
+    features['speed_range'] = speed_window.max() - speed_window.min()
+    features['speed_standard_dev'] = speed_window.std()
+    features['speed_energy'] = np.sum(speed_window ** 2)
+    features['speed_zcr'] = ((speed_window[:-1] * speed_window[1:]) < 0).sum() / len(speed_window)
+    features['speed_first_quartile'] = np.percentile(speed_window, 25)
+    features['speed_second_quartile'] = np.percentile(speed_window, 50)
+    features['speed_third_quartile'] = np.percentile(speed_window, 75)
+    features['speed_kfd'] = np.log10(len(speed_window)) / (np.log10(len(speed_window)) + np.log10(np.mean(np.abs(np.diff(speed_window)))))
+    features['speed_skewness'] = speed_window.skew()
+    features['speed_kurtosis'] = speed_window.kurtosis()
+    features['speed_sam_entropy'] = -np.sum(np.log2(speed_window / speed_window.sum()) * (speed_window / speed_window.sum()))
+    features['speed_shannon_entropy'] = -np.sum(speed_window * np.log2(speed_window + 1e-10))
+    features['speed_freq_variability'] = np.var(np.fft.fft(speed_window))
+    features['speed_spectral_entropy'] = -np.sum(np.log2(np.abs(np.fft.fft(speed_window)) + 1e-10))
+    features['speed_spectral_flux'] = np.sum(np.abs(np.diff(np.fft.fft(speed_window))))
+    features['speed_center_gravity_freq'] = np.sum(np.fft.fftfreq(len(speed_window)) * np.abs(np.fft.fft(speed_window)))
+    features['speed_dominant_freq'] = np.argmax(np.abs(np.fft.fft(speed_window)))
+    features['speed_avg_psd'] = np.mean(np.abs(np.fft.fft(speed_window))**2)
+    return features
+
+# 4つのインデックスを計算
+def calculate_feature_indices(features_df, label_df):
+    fisher_indices, correlation_indices, t_test_indices, mutual_info_indices = [], [], [], []
+    for column in features_df.columns:
+        group1 = features_df[column][label_df == 1]
+        group0 = features_df[column][label_df == 0]
+        fisher_index = ((group1.mean() - group0.mean()) ** 2) / (group1.var() + group0.var() + 1e-10)
+        fisher_indices.append(fisher_index)
+        correlation, _ = pearsonr(features_df[column], label_df)
+        correlation_indices.append(abs(correlation))
+        t_stat, _ = ttest_ind(group1, group0, equal_var=False)
+        t_test_indices.append(abs(t_stat))
+        mutual_info = mutual_info_classif(features_df[column].values.reshape(-1, 1), label_df)[0]
+        mutual_info_indices.append(mutual_info)
+    indices_df = pd.DataFrame({
+        'Feature': features_df.columns,
+        'Fisher': fisher_indices,
+        'Correlation': correlation_indices,
+        'T_test': t_test_indices,
+        'Mutual_Info': mutual_info_indices
+    })
+    return indices_df
+
+# ANFISによる特徴量選択とSVM最適化
+def calculate_importance_degree(params, indices_df):
+    weighted_scores = (indices_df['Fisher'] * params[0] +
+                       indices_df['Correlation'] * params[1] +
+                       indices_df['T_test'] * params[2] +
+                       indices_df['Mutual_Info'] * params[3])
+    importance_degree = np.where(weighted_scores > 0.75, 1, np.where(weighted_scores > 0.4, 0.5, 0))
+    return importance_degree
+
+def select_features_by_importance(importance_degree, features_df):
+    selected_features = features_df.loc[:, importance_degree == 1]
+    return selected_features
+
+def optimize_anfis_svm(X, y):
+    def objective(params):
+        anfis_params = params[:4]
+        C, gamma = params[4], params[5]
+        importance_degree = calculate_importance_degree(anfis_params, indices_df)
+        selected_features = select_features_by_importance(importance_degree, X)
+        if selected_features.shape[1] > 0:
+            svm_model = SVC(kernel='rbf', C=C, gamma=gamma)
+            svm_model.fit(selected_features, y)
+            train_accuracy = accuracy_score(y, svm_model.predict(selected_features))
+            return -train_accuracy
+        else:
+            return float('inf')
+
+    bounds = [(0, 1)] * 4 + [(0.1, 10), (0.001, 1)]
+    result = differential_evolution(objective, bounds, maxiter=10, popsize=3)
+    return result.x
+
+# データの読み込み
+file_path = './dataset/data_feat_Arefnezhad2019_all.csv'
+data_new = pd.read_csv(file_path)
+
+# ステアリング速度の計算
+time_diffs = data_new['row LSL time'].diff().dropna()
+sampling_rate = 1 / time_diffs.mean()
+data_new['Steering_Speed'] = data_new['Steering_Wheel_Pos'].diff().fillna(0) * sampling_rate
+
+# 3秒間の時間窓で特徴量を抽出
+window_size = int(3 * sampling_rate)
+num_windows = len(data_new) // window_size
+features_list = []
+
+for i in range(num_windows):
+    angle_window = data_new['Steering_Wheel_Pos'][i*window_size:(i+1)*window_size]
+    speed_window = data_new['Steering_Speed'][i*window_size:(i+1)*window_size]
+    features = calculate_features_for_signals(angle_window, speed_window)
+    features_list.append(features)
+features_df = pd.DataFrame(features_list)
+features_df['drowsy'] = data_new['drowsy'][:num_windows * window_size:window_size].reset_index(drop=True)
+
+
+# インデックスの計算とクリーンアップ
+label_df = features_df['drowsy']
+features_only_df = features_df.drop(columns=['drowsy']).replace([np.inf, -np.inf, np.nan], 0)
+indices_df = calculate_feature_indices(features_only_df, label_df)
+
+
+optimal_params = optimize_anfis_svm(features_only_df, label_df)
+best_anfis_params = optimal_params[:4]
+best_C, best_gamma = optimal_params[4], optimal_params[5]
+importance_degree = calculate_importance_degree(best_anfis_params, indices_df)
+selected_features_final = select_features_by_importance(importance_degree, features_only_df)
+
+# 訓練、検証、テストセットにデータを分割
+X_train, X_temp, y_train, y_temp = train_test_split(selected_features_final, label_df, test_size=0.4, random_state=42)
+X_val, X_test, y_val, y_test = train_test_split(X_temp, y_temp, test_size=0.5, random_state=42)
+
+# SVMモデルの構築と訓練
+svm_model_final = SVC(kernel='rbf', C=best_C, gamma=best_gamma)
+svm_model_final.fit(X_train, y_train)
+
+# 各セットでの精度計算
+train_accuracy = accuracy_score(y_train, svm_model_final.predict(X_train))
+val_accuracy = accuracy_score(y_val, svm_model_final.predict(X_val))
+test_accuracy = accuracy_score(y_test, svm_model_final.predict(X_test))
+
+# 結果の表示
+{
+    "Train Accuracy": train_accuracy,
+    "Validation Accuracy": val_accuracy,
+    "Test Accuracy": test_accuracy,
+}
+
+# 結果の表示
+print(
+        "results",
+        "\nBest ANFIS Parameters (Weights):", best_anfis_params,
+        "\nBest SVM Parameters (C, gamma) :", (best_C, best_gamma),
+        "\nTrain Accuracy                 :", train_accuracy,
+        "\nValidation Accuracy            :", val_accuracy,
+        "\nTest Accuracy                  :", test_accuracy,
+)
+
+print("selected features\n",selected_features_final.columns.tolist())
+
+# 重要度インデックスの長さを調整して再実行します
+
+def select_features_by_importance_adjusted(importance_degree, features_df):
+    # データフレームの列に合わせて、重要度インデックスを調整
+    importance_degree_adjusted = importance_degree[:features_df.shape[1]]
+    selected_features = features_df.loc[:, importance_degree_adjusted == 1]
+    return selected_features
+
+# 最適化関数内での特徴量選択部分を修正して実行します
+def optimize_anfis_svm_for_val_accuracy_adjusted(X_train, y_train, X_val, y_val):
+    def objective(params):
+        anfis_params = params[:4]
+        C, gamma = params[4], params[5]
+
+        # インデックスに基づく重要度を計算し、出力が1の特徴量を選択
+        importance_degree = calculate_importance_degree(anfis_params, indices_df)
+        selected_features_train = select_features_by_importance_adjusted(importance_degree, X_train)
+        selected_features_val = select_features_by_importance_adjusted(importance_degree, X_val)
+
+        if selected_features_train.shape[1] > 0 and selected_features_val.shape[1] > 0:
+            svm_model = SVC(kernel='rbf', C=C, gamma=gamma)
+            svm_model.fit(selected_features_train, y_train)
+            val_accuracy = accuracy_score(y_val, svm_model.predict(selected_features_val))
+            return -val_accuracy  # 検証精度最大化のため符号反転
+        else:
+            return float('inf')  # 特徴量がない場合、大きな値で無効にする
+
+    # PSOのパラメータ範囲
+    bounds = [(0, 1)] * 4 + [(0.1, 10), (0.001, 1)]
+
+    # PSOの実行
+    result = differential_evolution(objective, bounds, maxiter=10, popsize=3)
+    return result.x
+
+# PSOを用いた最適化
+def optimize_anfis_svm_with_pso(X_train, y_train, X_val, y_val):
+    def objective(params):
+        anfis_params = params[:4]
+        C, gamma = params[4], params[5]
+        importance_degree = calculate_importance_degree(anfis_params, indices_df)
+        selected_features_train = select_features_by_importance(importance_degree, X_train)
+        selected_features_val = select_features_by_importance(importance_degree, X_val)
+
+        if selected_features_train.shape[1] > 0 and selected_features_val.shape[1] > 0:
+            svm_model = SVC(kernel='rbf', C=C, gamma=gamma)
+            svm_model.fit(selected_features_train, y_train)
+            val_accuracy = accuracy_score(y_val, svm_model.predict(selected_features_val))
+            return -val_accuracy
+        else:
+            return float('inf')
+
+    # パラメータの範囲設定
+    lb = [0, 0, 0, 0, 0.1, 0.001]
+    ub = [1, 1, 1, 1, 10, 1]
+
+    # PSOの実行
+    best_params, _ = pso(objective, lb, ub, swarmsize=50, maxiter=100)
+    return best_params
+
+# 訓練と検証データに対する最適化実行
+optimal_params_val_accuracy_adjusted = optimize_anfis_svm_for_val_accuracy_adjusted(X_train, y_train, X_val, y_val)
+
+# 最適化されたパラメータでANFISとSVMを設定
+best_anfis_params_val_adj = optimal_params_val_accuracy_adjusted[:4]
+best_C_val_adj, best_gamma_val_adj = optimal_params_val_accuracy_adjusted[4], optimal_params_val_accuracy_adjusted[5]
+importance_degree_val_adj = calculate_importance_degree(best_anfis_params_val_adj, indices_df)
+
+# 選択された特徴量で再度評価
+selected_features_train_val_adj = select_features_by_importance_adjusted(importance_degree_val_adj, X_train)
+selected_features_val_val_adj = select_features_by_importance_adjusted(importance_degree_val_adj, X_val)
+selected_features_test_val_adj = select_features_by_importance_adjusted(importance_degree_val_adj, X_test)
+
+# SVMモデルの訓練と評価
+svm_model_final_val_adj = SVC(kernel='rbf', C=best_C_val_adj, gamma=best_gamma_val_adj)
+svm_model_final_val_adj.fit(selected_features_train_val_adj, y_train)
+
+# 精度の計算
+train_accuracy_val_adj = accuracy_score(y_train, svm_model_final_val_adj.predict(selected_features_train_val_adj))
+val_accuracy_val_adj = accuracy_score(y_val, svm_model_final_val_adj.predict(selected_features_val_val_adj))
+test_accuracy_val_adj = accuracy_score(y_test, svm_model_final_val_adj.predict(selected_features_test_val_adj))
+
+# 結果の表示
+print(
+"result (opt for val acc)",
+"\nBest ANFIS Parameters (Weights for Fisher, Correlation, T-test, Mutual Info):", best_anfis_params_val_adj,
+"\nBest SVM Parameters (C, gamma)                                              :", (best_C_val_adj, best_gamma_val_adj),
+"\nTrain Accuracy                                                              :", train_accuracy_val_adj,
+"\nValidation Accuracy                                                         :", val_accuracy_val_adj,
+"\nTest Accuracy                                                               :", test_accuracy_val_adj,
+)
+
+print("selected features\n",selected_features_final.columns.tolist())
