+++ project/archive/tests/related_work/arefnezhad2019/tags
@@ -1,0 +1,34 @@
+!_TAG_FILE_FORMAT	2	/extended format; --format=1 will not append ;" to lines/
+!_TAG_FILE_SORTED	1	/0=unsorted, 1=sorted, 2=foldcase/
+!_TAG_PROGRAM_AUTHOR	Darren Hiebert	/dhiebert@users.sourceforge.net/
+!_TAG_PROGRAM_NAME	Exuberant Ctags	//
+!_TAG_PROGRAM_URL	http://ctags.sourceforge.net	/official site/
+!_TAG_PROGRAM_VERSION	5.7	//
+calculate_feature_indices	.\ddd_Aref2019.py	/^def calculate_feature_indices(features_df, label_df):$/;"	f
+calculate_feature_indices	.\ddd_Aref2019_PSO.py	/^def calculate_feature_indices(features_df, label_df):$/;"	f
+calculate_feature_indices	.\ddd_Aref2019_PSO_rev1.py	/^def calculate_feature_indices(features_df, label_df):$/;"	f
+calculate_features_for_signals	.\ddd_Aref2019.py	/^def calculate_features_for_signals(angle_window, speed_window):$/;"	f
+calculate_features_for_signals	.\ddd_Aref2019_PSO.py	/^def calculate_features_for_signals(angle_window, speed_window):$/;"	f
+calculate_features_for_signals	.\ddd_Aref2019_PSO_rev1.py	/^def calculate_features_for_signals(angle_window, speed_window):$/;"	f
+calculate_importance_degree	.\ddd_Aref2019.py	/^def calculate_importance_degree(params, indices_df):$/;"	f
+calculate_importance_degree	.\ddd_Aref2019_PSO.py	/^def calculate_importance_degree(params, indices_df):$/;"	f
+calculate_importance_degree	.\ddd_Aref2019_PSO_rev1.py	/^def calculate_importance_degree(params, indices_df):$/;"	f
+calculate_metrics	.\ddd_Aref2019_PSO_rev1.py	/^def calculate_metrics(y_true, y_pred):$/;"	f
+ification	.\ddd_Aref2019_PSO_rev1.py	/^    precision = precision_score(y_true, y_pred, average='binary')  # 'binary' for binary classification$/;"	c	function:calculate_metrics
+objective	.\ddd_Aref2019.py	/^    def objective(params):$/;"	f	function:optimize_anfis_svm
+objective	.\ddd_Aref2019.py	/^    def objective(params):$/;"	f	function:optimize_anfis_svm_for_val_accuracy_adjusted
+objective	.\ddd_Aref2019.py	/^    def objective(params):$/;"	f	function:optimize_anfis_svm_with_pso
+objective	.\ddd_Aref2019_PSO.py	/^    def objective(params):$/;"	f	function:optimize_anfis_svm_with_pso
+objective	.\ddd_Aref2019_PSO_rev1.py	/^    def objective(params):$/;"	f	function:optimize_anfis_svm_with_pso
+optimize_anfis_svm	.\ddd_Aref2019.py	/^def optimize_anfis_svm(X, y):$/;"	f
+optimize_anfis_svm_for_val_accuracy_adjusted	.\ddd_Aref2019.py	/^def optimize_anfis_svm_for_val_accuracy_adjusted(X_train, y_train, X_val, y_val):$/;"	f
+optimize_anfis_svm_with_pso	.\ddd_Aref2019.py	/^def optimize_anfis_svm_with_pso(X_train, y_train, X_val, y_val):$/;"	f
+optimize_anfis_svm_with_pso	.\ddd_Aref2019_PSO.py	/^def optimize_anfis_svm_with_pso(X_train, y_train, X_val, y_val):$/;"	f
+optimize_anfis_svm_with_pso	.\ddd_Aref2019_PSO_rev1.py	/^def optimize_anfis_svm_with_pso(X_train, y_train, X_val, y_val):$/;"	f
+plot_roc_from_csv	.\plot_roc.py	/^def plot_roc_from_csv(file_name, label):$/;"	f
+save_roc_data	.\ddd_Aref2019_PSO.py	/^def save_roc_data(file_name, y_true, y_scores):$/;"	f
+save_roc_data	.\ddd_Aref2019_PSO_rev1.py	/^def save_roc_data(file_name, y_true, y_scores):$/;"	f
+select_features_by_importance	.\ddd_Aref2019.py	/^def select_features_by_importance(importance_degree, features_df):$/;"	f
+select_features_by_importance	.\ddd_Aref2019_PSO.py	/^def select_features_by_importance(importance_degree, features_df):$/;"	f
+select_features_by_importance	.\ddd_Aref2019_PSO_rev1.py	/^def select_features_by_importance(importance_degree, features_df):$/;"	f
+select_features_by_importance_adjusted	.\ddd_Aref2019.py	/^def select_features_by_importance_adjusted(importance_degree, features_df):$/;"	f
