--- project/tests/related_work/arefnezhad2019/output.txt
@@ -1,23 +1,0 @@
-Stopping search: maximum iterations reached --> 100
-['angle_range', 'angle_standard_dev', 'angle_energy', 'angle_first_quartile', 'angle_second_quartile', 'angle_third_quartile', 'angle_shannon_entropy', 'angle_freq_variability', 'angle_spectral_flux', 'angle_dominant_freq', 'angle_avg_psd', 'speed_range', 'speed_standard_dev', 'speed_energy', 'speed_first_quartile', 'speed_third_quartile', 'speed_freq_variability', 'speed_spectral_flux', 'speed_avg_psd']
-['angle_range', 'angle_standard_dev', 'angle_energy', 'angle_first_quartile', 'angle_second_quartile', 'angle_third_quartile', 'angle_shannon_entropy', 'angle_freq_variability', 'angle_spectral_flux', 'angle_dominant_freq', 'angle_avg_psd', 'speed_range', 'speed_standard_dev', 'speed_energy', 'speed_first_quartile', 'speed_third_quartile', 'speed_freq_variability', 'speed_spectral_flux', 'speed_avg_psd']
-['angle_range', 'angle_standard_dev', 'angle_energy', 'angle_first_quartile', 'angle_second_quartile', 'angle_third_quartile', 'angle_shannon_entropy', 'angle_freq_variability', 'angle_spectral_flux', 'angle_dominant_freq', 'angle_avg_psd', 'speed_range', 'speed_standard_dev', 'speed_energy', 'speed_first_quartile', 'speed_third_quartile', 'speed_freq_variability', 'speed_spectral_flux', 'speed_avg_psd']
-train_conf
- [[897 697]
- [832 817]]
-val_conf
- [[329 241]
- [255 256]]
-test_conf
- [[279 259]
- [251 292]]
-AUC_train   : 0.529
-AUC_val     : 0.539
-AUC_test    : 0.528
-results
-Best ANFIS Parameters (Weights): [1.         0.         0.69572227 0.66017369]
-Best SVM Parameters (C, gamma) : (7.249377996495855, 1.0)
-Train Accuracy                 : 0.528522972556275
-Validation Accuracy            : 0.5411655874190564
-Test Accuracy                  : 0.5282146160962072
-elaped time: 5883.185302734375
